partial fix for #99
I think the code is faster since the latests commits
